+++ src/func/constants.py
@@ -1,4 +1,4 @@
-# TODO TODO refactor this into src/utils.py
+# TODO TODO refactor this into src/general.py
 version = "5.1.0"
 
 header = {
